requirements.txt
streamlit
rembg
pillow
requests
numpy
onnxruntime